Renamed mock files and integrated person to use them
