+++ lib/test/poll.test.ts
@@ -6,8 +6,8 @@ import { ObjectId } from "mongodb";
 import app from "../App"
 import { PollSchema } from '../models/PollModel';
 import { PersonSchema } from '../models/PersonModel';
-import { people } from './person_test_objects';
-import { open_poll_id, closed_poll_id, closed_poll, halfway_poll_id, halfway_poll, polls } from './poll_test_objects';
+import { people } from './person_mock';
+import { open_poll_id, closed_poll_id, closed_poll, halfway_poll_id, halfway_poll, polls } from './poll_mock';
 
 
 const Poll = mongoose.model('Poll', PollSchema);
